# feat(bitnet): implement scaled dot-product attention