--- a/.gitignore
+++ b/.gitignore
@@ -15,3 +15,5 @@ tensor.test
 
 # BitNet model files
 pkg/bitnet/internal/assets/models/
+
+math.test